refactor: split dom.js into focused sub-modules (dom-core, dom-buttons, dom-tabs, dom-lists)
Break the dom.js hotspot (50 dependents) into 4 cohesive sub-modules:
- dom-core.js: _el, _vis (primitives)
- dom-buttons.js: createActionButton, renderButtonBar, buildDomainButtonBar
- dom-tabs.js: buildTabButton, buildTabBar
- dom-lists.js: renderList, buildChevronRow

dom.js becomes a barrel re-export for full backward-compatibility.

Closes #541
